Replaced smif.outputs with reference to smif.inputs
