+++ smif/inputs.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-"""Encapsulates the inputs to a sector model
-
+"""Encapsulates the input or output parameters of a sector model,
+for example::
 
         - name: petrol_price
           spatial_resolution: GB
@@ -102,11 +102,27 @@ class ModelInputs(object):
         A dictionary of key: val pairs including a list of input types and
         names, followed by nested dictionaries of input attributes
     """
-    def __init__(self, inputs):
-        if 'dependencies' not in inputs:
-            inputs['dependencies'] = []
+    def __init__(self, parameters):
+        if 'dependencies' not in parameters:
+            parameters['dependencies'] = []
 
-        self._parameters = ParameterList(inputs['dependencies'])
+        self._parameters = ParameterList(parameters['dependencies'])
+
+        if 'metrics' not in parameters:
+            parameters['metrics'] = []
+
+        self._metrics = ParameterList(parameters['metrics'])
+        self.logger = logging.getLogger(__name__)
+
+    @property
+    def metrics(self):
+        """A list of the model result metrics
+
+        Returns
+        =======
+        :class:`smif.inputs.ParameterList`
+        """
+        return self._metrics
 
     @property
     def dependencies(self):
@@ -120,3 +136,29 @@ class ModelInputs(object):
 
     def __len__(self):
         return len(self.dependencies)
+
+    def get_spatial_res(self, name):
+        for metric in self._metrics:
+            if metric.name == name:
+                spatial_resolution = metric.spatial_resolution
+                break
+        else:
+            raise ValueError("No output found for name {}".format(name))
+        return spatial_resolution
+
+    def get_temporal_res(self, name):
+        for metric in self._metrics:
+            if metric.name == name:
+                temporal_resolution = metric.temporal_resolution
+                break
+        else:
+            raise ValueError("No output found for name '{}'".format(name))
+        return temporal_resolution
+
+    @property
+    def spatial_resolutions(self):
+        return self._metrics.spatial_resolutions
+
+    @property
+    def temporal_resolutions(self):
+        return self._metrics.temporal_resolutions
